requirements.txt
absl-py
chex
dm-haiku
dm-tree
git+https://github.com/deepmind/einshape
jax
numpy
optax
tqdm
typing_extensions
